requirements.txt
pandas==2.2.2
openpyxl==3.1.2
supabase==2.7.0
python-dotenv==1.0.1
pytest==8.2.0
